Adds automatic function for phase seperation
Adds automatic function to find optimum bin number for phase separation.
Added function is plane_count to determine the physically useful frequency of alloy phase peaks based on the number of planes which are fitted. In addition, the highest number of  peaks in one plane is used as a base for the minimal number of alloy phases to determine.