New version of the Notebook with updated examples
All used functions are transferred into a module to clean up the notebook. Furthermore, The plot function are cleaned up and appearance is improved by adding titles to the legend. The ox_index function is improved by checking the validity of the uncertainty range and correcting it to lower values if needed. This avoids, the effect of overlapping and therefore not considered values.